+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Delete Inventory and MARC holdings records when record Source = MARC. Refs UIIN-1559.
 * Increment `stripes` to `v7`, `react` to `v17`. Refs UIIN-1564.
 * Edit MARC Holdings via quickMARC. Refs UIIN-1636.
+* Make mapped fields read-only for MARC holdings records. Refs UIIN-1639.
 
 ## [7.1.4](https://github.com/folio-org/ui-inventory/tree/v7.1.4) (2021-08-05)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v7.1.3...v7.1.4)